Server/requirements.txt
aniso8601==8.0.0
autopep8==1.5.4
bcrypt==3.1.4
blinker==1.4
businesstimedelta==1.0.1
certifi==2016.2.28
cffi==1.11.5
chardet==3.0.4
click==6.7
convertdate==2.3.0
cryptography==3.2.1
Flask==1.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
holidays==0.10.3
idna==2.10
itsdangerous==0.24
Jinja2==2.10
korean-lunar-calendar==0.2.1
MarkupSafe==1.0
Pillow==5.3.0
pycodestyle==2.6.0
pycparser==2.18
PyJWT==1.7.1
PyMeeus==0.3.7
pyodbc==4.0.30
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.4
requests==2.25.0
six==1.11.0
SQLAlchemy==1.2.7
toml==0.10.2
urllib3==1.26.2
Werkzeug==0.14.1
WTForms==2.1
